+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.4.0
+
+* Support ModelSafe's new lazy loading association declaration feature
+* Add `save` operation to querying that either updates or creates a model instance
+  automatically based off the primary key of the model
+
 # 0.3.1
 
 * Re-add `getModelPrimary` as `getInternalModePrimary` to the database class